+++ services/domainservices/mgmt/2020-01-01/aad/CHANGELOG.md
@@ -1,2 +1,18 @@
-# Change History
+# Unreleased
 
+## Additive Changes
+
+### New Funcs
+
+1. MigrationProperties.MarshalJSON() ([]byte, error)
+
+### Struct Changes
+
+#### New Structs
+
+1. MigrationProgress
+1. MigrationProperties
+
+#### New Struct Fields
+
+1. DomainServiceProperties.MigrationProperties